I FIXED IT I'M THE GREATEST. Two hours wasted...
Details:
Certain aspects of the database had changed since the root file of our project was added to git. The models and everything needed to be updated. The update wizard failed. I had to manually resolve the conflicts...

It's all good now. Now that the errors are handled, I just have to finish implemented the employee authentication page.