Log in now works. Not fully implemented.
Currently if you type in a user name and password that is in the DB a log in notice will present you with your first name and last name.
Need to implement changing view for if log in fails.
Need to implement some way of keeping this login flagged for both user ID as well as privileges.  This will be the hardest part. I'm not sure how to even start that.